refactor(api-compare,activerecord): fold the narrow call ratchet into the wide one, drop the encryption previous-types memo, empty the SchemaStatements Omit
RFC 0084: the narrow RFC 0044 call ratchet was strictly subsumed by
population by the wide RFC 0047 one, at the cost of a duplicate artifact, a
second CI step, a second reseed command and the two-artifact
API_COMPARE_FORCE trap. Retired: the narrow baseline (14 rows), CI step,
`api:calls` / `api:calls:reseed` scripts, SIGNIFICANT_CALLS and
narrowCallsApplies. `compare.ts` now computes call sets — and writes an
artifact — only under `--wide-calls`. The shared machinery moves to the
neutral `call-mismatch-baseline.ts`.

Of the 14 narrow rows, 4 are flagged by the wide population and were already
baselined there carrying the seeded placeholder; their reviewed RFC 0044
reasons replace it (marks tightened 12→11 / 19→17 / 33→32). The other 10 are
not flagged by the wide gate at all, so nothing is lost by dropping them.

RFC 0072: EncryptedAttributeType's generation counter and its three memos
(`previousTypes`, `_effectivePreviousSchemes`, `isSerializeWithOldest`) are a
trails invention. Rails memoizes with a plain `||=` because a scheme's
previous_schemes is frozen at `encrypts` time (encryptable_record.rb:73);
trails resolves the global previous schemes lazily, so the value is not
immutable and a memo pins a pre-`configure` answer. Recompute per call, and
route the two `previous_types_without_clean_text` call sites through it as
encrypted_attribute_type.rb:95,123 does.

RFC 0051: `addColumns` was declared on AbstractAdapter with a per-column
record vararg, which is not Rails' `add_columns(table_name, *column_names,
type:, **options)` (schema_statements.rb:643) and was the only real conflict
left in the SchemaStatements merged interface's `Omit`. Converged, and the
`Omit` is gone entirely — the enum / currentDatabase entries no longer
conflict.

Closes-story: fold-narrow-call-ratchet-into-wide
Closes-story: remove-encrypted-attribute-type-previous-types-memo
Closes-story: schema-statements-merged-interface-omits-conflicting-adapter-members
