+++ package.json
@@ -27,8 +27,6 @@
     "api:build": "pnpm tsx scripts/api-compare/build.ts",
     "api:reasons": "pnpm tsx scripts/api-compare/lint-missing-rails-call-reasons.ts",
     "api:detached": "pnpm tsx scripts/api-compare/lint-detached-jsdoc-tags.ts",
-    "api:calls": "pnpm tsx scripts/api-compare/lint-call-mismatches.ts",
-    "api:calls:reseed": "API_COMPARE_FORCE=1 pnpm api:compare && pnpm tsx scripts/api-compare/lint-call-mismatches.ts --write --no-regen",
     "api:calls:wide": "pnpm tsx scripts/api-compare/lint-call-mismatches-wide.ts",
     "api:calls:wide:report": "pnpm tsx scripts/api-compare/lint-call-mismatches-wide.ts --report",
     "api:calls:wide:unreviewed": "pnpm tsx scripts/api-compare/lint-call-mismatches-wide.ts --unreviewed",
