+++ packages/activerecord/src/connection-adapters/abstract-adapter.ts
@@ -259,9 +259,12 @@ export interface AbstractAdapter {
     allowNull: boolean,
     defaultValue?: unknown,
   ): Promise<void>;
+  // Rails' `add_columns(table_name, *column_names, type:, **options)`
+  // (schema_statements.rb:643): the trailing hash carries `type` plus the
+  // per-column options, so it is the last vararg, not a per-column record.
   addColumns(
     tableName: string,
-    ...columns: Array<{ name: string; type: ColumnType; options?: ColumnOptions }>
+    ...args: [...string[], { type: ColumnType } & ColumnOptions]
   ): Promise<void>;
   removeColumn(
     tableName: string,
