+++ scripts/api-compare/call-mismatches-wide-exclude/activerecord/base.json
@@ -102,7 +102,7 @@
     "tsFile": "base.ts",
     "rubyName": "update",
     "call": "save",
-    "reason": "Baseline (RFC 0047): wide call-set flag seeded when the wide ratchet landed; bucket (b) equivalent or (c) noise pending per-cluster burndown review."
+    "reason": "Confirmed equivalent (RFC 0044, update-delegates-to-save-cluster): base.ts's class-level `update` mirrors Rails `Persistence::ClassMethods#update` / `Relation#update` (persistence.rb:132, relation.rb:623), which never call `save` directly -- they delegate per record to the instance `#update`. base.ts does the same via `record.update(a)` (base.ts:454); the `save` lands in persistence.ts instance `update` (`self.save()`), which is itself unflagged. Indirect delegation, not an omission."
   },
   {
     "package": "activerecord",
@@ -123,7 +123,7 @@
     "tsFile": "base.ts",
     "rubyName": "update!",
     "call": "save!",
-    "reason": "Baseline (RFC 0047): wide call-set flag seeded when the wide ratchet landed; bucket (b) equivalent or (c) noise pending per-cluster burndown review."
+    "reason": "Confirmed equivalent (RFC 0044, update-delegates-to-save-cluster): base.ts's class-level `update!` mirrors Rails `Persistence::ClassMethods#update!` / `Relation#update!` (persistence.rb:158, relation.rb:631), which delegate per record to the instance `#update!` rather than calling `save!` directly. base.ts does the same via `record.updateBang(a)` (base.ts:453); the `save!` lands in persistence.ts `updateBang` (`self.saveBang()`). Indirect delegation, not an omission."
   },
   {
     "package": "activerecord",
