+++ scripts/api-compare/call-mismatches-wide-unreviewed/activerecord/associations/collection-association.json
@@ -1,3 +1,3 @@
 {
-  "max": 12
+  "max": 11
 }
