+++ scripts/api-compare/call-mismatches-wide-unreviewed/activerecord/base.json
@@ -1,3 +1,3 @@
 {
-  "max": 19
+  "max": 17
 }
