+++ scripts/api-compare/call-mismatches-wide-unreviewed/activerecord/migration.json
@@ -1,3 +1,3 @@
 {
-  "max": 33
+  "max": 32
 }
