+++ scripts/prism-codegen/score-cli.ts
@@ -82,14 +82,9 @@ async function railsLibFiles(): Promise<{ path: string; source: string }[]> {
   return out;
 }
 
-// The deviation catalog's call half: the flat exclude list plus every file of
-// the split wide exclude tree. Both share the entry shape, so the scorer sees
-// one merged list.
+// The deviation catalog's call half: every file of the split wide exclude tree.
 async function loadExcludes(): Promise<ExcludeEntry[]> {
-  const files = [
-    path.join(API_COMPARE, "call-mismatches-exclude.json"),
-    ...(await listJsonFiles(path.join(API_COMPARE, "call-mismatches-wide-exclude"))),
-  ];
+  const files = await listJsonFiles(path.join(API_COMPARE, "call-mismatches-wide-exclude"));
   const out: ExcludeEntry[] = [];
   for (const file of files) {
     let source;
